Moves article tests
